requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
ta>=0.11.0
python-dotenv>=1.0.0
pydantic>=2.0.0
flask>=3.0.0
apscheduler>=3.10.0
pytest>=7.0.0
